Update README.md
Can we consider replacing the image of Using the chrome extension by just capturing it only from Tutorials Point heading as there are other  browser tabs also displayed.
I the same section point 2 the image seems to be missing.
Can we have the Developer notes section before the installation steps as it talks about requirements and packaging section information.
The Packaging section reasons the font is  different. I tried to but could not do it. Can you please check.